[Tushar] add: Edit the Contact details

+++ AddressBook/Contact.cs
@@ -14,7 +14,7 @@ namespace AddressBook
 
         public string Address { get; set; }
         public string City { get; set; }
-        public long Zip { get; set; }
+        public int Zip { get; set; }
         public string State { get; set; }
         public string Email { get; set; }
         public string Phone { get; set; }
